BrainPlowProjects/BackendGigsGenie/requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.22.0
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
cffi==1.10.0
channels==1.1.8
click==6.7
constantly==15.1.0
cryptography==2.0.3
daphne==1.3.0
dj-database-url==0.4.2
Django==1.9
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
enum34==1.1.6
Flask==0.12.2
gunicorn==19.7.1
haikunator==1.0.1
hyperlink==17.3.1
idna==2.2
incremental==17.5.0
ipaddress==1.0.17
itsdangerous==0.24
jieba==0.38
Jinja2==2.9.6
keyring==10.3.1
keyrings.alt==2.2
legacy==0.1.6
lockfile==0.12.2
Markdown==2.6.8
MarkupSafe==1.0
msgpack-python==0.4.7
ndg-httpsclient==0.4.2
numpy==1.12.1
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
psycopg2==2.7.1
py2neo==3.1.2
pyasn1==0.1.9
pycparser==2.18
pyOpenSSL==17.2.0
pypinyin==0.22.0
python-cloudfiles==1.7.10
pytz==2017.2
pyxdg==0.25
redis==2.10.5
requests==2.10.0
SecretStorage==2.3.1
six==1.10.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.15.1
virtualenv==15.1.0
websocket-client==0.44.0
Werkzeug==0.12.2
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0
zope.interface==4.4.3
django-extensions
pygraphviz==1.3.1
pyasn1-modules==0.1.4
service-identity==17.0.0
djangorestframework-jwt==1.11.0